moved EnableNotifications and Onboarding into screens/onboarding folder
